Bindings: no more memory errors, fix typing.

+++ zig-impl/libipc.h
@@ -14,11 +14,11 @@ enum event_types {
 };
 
 int ipc_context_init (void** ptr);
-int ipc_service_init (void* ctx, int* servicefd, const char* service_name, size_t service_name_len);
-int ipc_connect_service (void* ctx, int* servicefd, const char* service_name, size_t service_name_len);
+int ipc_service_init (void* ctx, int* servicefd, const char* service_name, unsigned short service_name_len);
+int ipc_connect_service (void* ctx, int* servicefd, const char* service_name, unsigned short service_name_len);
 void ipc_context_deinit (void* ctx);
-int ipc_write (void* ctx, int servicefd, char* mcontent, size_t mlen);
-int ipc_schedule (void* ctx, int servicefd, const char* mcontent, size_t mlen);
+int ipc_write (void* ctx, int servicefd, char* mcontent, unsigned int mlen);
+int ipc_schedule (void* ctx, int servicefd, const char* mcontent, unsigned int mlen);
 int ipc_read_fd (void* ctx, int fd, char* buffer, size_t* buflen);
 int ipc_read (void* ctx, size_t index, char* buffer, size_t* buflen);
 int ipc_wait_event(void* ctx, char* t, size_t* index, int* originfd, char* buffer, size_t* buflen);
